+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/Board/dto/request/PostSubCommentRequestDto.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class PostSubCommentRequestDto {
     private String content;
-    private Integer parent_comment_id;
+    private Long parent_comment_id;
 }
